🔥 chore: move components and utils to the src folder
