+++ src/components/AccordionBuildingLog.tsx
