+++ src/types/environment.d.ts
